# authz-svc/api/openapi.yaml
openapi: 3.0.3
info:
  title: Authorization Service API
  version: 1.0.0
paths:
  /v1/check:
    post:
      operationId: checkPermission
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CheckRequest'
      responses:
        '200':
          description: Permission decision
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CheckResponse'
        '400':
          $ref: '#/components/responses/Error'
        '503':
          $ref: '#/components/responses/Error'
  /v1/organizations/{organizationId}/users/{userId}/roles/admin:
    put:
      operationId: assignAdmin
      parameters:
        - $ref: '#/components/parameters/ActorId'
        - $ref: '#/components/parameters/OrganizationId'
        - $ref: '#/components/parameters/UserId'
      responses:
        '204': { description: Admin role assigned }
        '400': { $ref: '#/components/responses/Error' }
        '403': { $ref: '#/components/responses/Error' }
        '503': { $ref: '#/components/responses/Error' }
    delete:
      operationId: revokeAdmin
      parameters:
        - $ref: '#/components/parameters/ActorId'
        - $ref: '#/components/parameters/OrganizationId'
        - $ref: '#/components/parameters/UserId'
      responses:
        '204': { description: Admin role revoked }
        '400': { $ref: '#/components/responses/Error' }
        '403': { $ref: '#/components/responses/Error' }
        '503': { $ref: '#/components/responses/Error' }
components:
  parameters:
    ActorId:
      name: X-User-ID
      in: header
      required: true
      schema: { type: string, format: uuid }
    OrganizationId:
      name: organizationId
      in: path
      required: true
      schema: { type: string, format: uuid }
    UserId:
      name: userId
      in: path
      required: true
      schema: { type: string, format: uuid }
  responses:
    Error:
      description: Error response
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
  schemas:
    CheckRequest:
      type: object
      required: [subjectId, organizationId, permission]
      properties:
        subjectId: { type: string, format: uuid }
        organizationId: { type: string, format: uuid }
        permission: { type: string, enum: [list_users, create_user] }
    CheckResponse:
      type: object
      required: [allowed]
      properties:
        allowed: { type: boolean }
    Error:
      type: object
      required: [code, message]
      properties:
        code: { type: string }
        message: { type: string }